% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_finance.R
\name{CleanStockPrices}
\alias{CleanStockPrices}
\title{Cleans raw stock prices fetched from Yahoo (quantmod) API.}
\usage{
CleanStockPrices(stock_prices)
}
\description{
Cleans raw stock prices fetched from Yahoo (quantmod) API.
}
