% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_finance.R
\name{GetDataForTicker}
\alias{GetDataForTicker}
\title{Downloads data for each ticker.}
\usage{
GetDataForTicker(path_prefix, ticker, beg_date, end_date)
}
\description{
Downloads data for each ticker.
}
