% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_finance.R
\name{CalculateReturns}
\alias{CalculateReturns}
\title{Calculates percent returns for short and long term trades.}
\usage{
CalculateReturns(stock_prices, congressional_trades)
}
\description{
Calculates percent returns for short and long term trades.
}
